pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "tabbycatimporter"
version = "0.0.1"
dependencies = [
    "click",
    "pyyaml",
    "requests",
]

[project.scripts]
tbimport = "tbimport.cli:cli"